dww_backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.12
    # via
    #   aiohttp-retry
    #   twilio
aiohttp-retry==2.9.1
    # via twilio
aiosignal==1.3.2
    # via aiohttp
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
awsebcli==3.21.0
    # via -r requirements.in
botocore==1.35.99
    # via awsebcli
build==1.2.2.post1
    # via pip-tools
cement==2.10.14
    # via awsebcli
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via pip-tools
colorama==0.4.6
    # via
    #   awsebcli
    #   build
    #   click
django==5.1.4
    # via
    #   django-cors-headers
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.6.0
    # via -r requirements.in
django-phonenumber-field==8.0.0
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via djangorestframework-simplejwt
djangorestframework-simplejwt==5.4.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
jmespath==1.0.1
    # via botocore
multidict==6.3.0
    # via
    #   aiohttp
    #   yarl
mysqlclient==2.2.7
    # via -r requirements.in
packaging==24.2
    # via
    #   build
    #   pipdeptree
pathspec==0.10.1
    # via awsebcli
phonenumberslite==8.13.53
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
pipdeptree==2.25.0
    # via -r requirements.in
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via -r requirements.in
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   twilio
pypiwin32==223 ; sys_platform == "win32"
    # via
    #   -r requirements.in
    #   awsebcli
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   awsebcli
    #   botocore
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.2
    # via -r requirements.in
pywin32==308 ; sys_platform == "win32"
    # via
    #   -r requirements.in
    #   pypiwin32
pyyaml==6.0.2
    # via awsebcli
requests==2.32.3
    # via
    #   awsebcli
    #   twilio
semantic-version==2.10.0
    # via awsebcli
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
termcolor==2.5.0
    # via awsebcli
tomli==2.2.1
    # via
    #   build
    #   pip-tools
twilio==9.5.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   asgiref
    #   multidict
tzdata==2024.2
    # via django
urllib3==1.26.20
    # via
    #   awsebcli
    #   botocore
    #   requests
wcwidth==0.2.13
    # via awsebcli
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
